requirements.txt
flask
pytube
azure-storage-blob
python-dotenv